Revert "libsepol/cil: raise default attrs_expand_size to 2"
This reverts commit 770a59c445fd94975e092db0420b782fe6b78543.

Other parts of the SELinux userspace depend on certain attributes,
such as node_type, exisiting and this change breaks those parts.

Before this patch can be reapplied, we need to identify the attributes
that must never be expanded and create a CIL module with the needed
expandtypeattribute statements (or something similar).

Signed-off-by: James Carter <jwcarter@gmail.com>